requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic[email]==2.10.4
python-multipart==0.0.20
redis==5.0.1
aioredis==2.0.1
celery==5.3.6
pillow==10.2.0
aiofiles==23.2.1
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.27.0
pytest-cov==4.1.0